Retry the UNAVAILABLE cases for YDB
