+++ internal/static/embed.go
@@ -20,6 +20,7 @@ const (
 	StroppyDTSFileName FileName = "stroppy.d.ts"
 
 	HelpersFileName FileName = "helpers.ts"
+	RetryFileName   FileName = "retry.ts"
 	DatagenFileName FileName = "datagen.ts"
 
 	ParseSQLTSFileName FileName = "parse_sql.ts"
@@ -36,6 +37,7 @@ const (
 var StaticFiles = []FileName{
 	ProtoJSFileName,
 	HelpersFileName,
+	RetryFileName,
 	DatagenFileName,
 	ParseSQLJSFileName,
 	StroppyDTSFileName,
